refactor: 인증 로직과 비즈니스 로직 분리에 따른 import 및 코드 구조 수정
