+++ src/main/java/com/yanolja/areas/auth/controller/AuthController.java
@@ -1,6 +1,6 @@
 package com.yanolja.areas.auth.controller;
 
-import com.yanolja.areas.auth.domain.User;
+import com.yanolja.areas.user.domain.User;
 import com.yanolja.areas.auth.dto.LoginRequest;
 import com.yanolja.areas.auth.dto.LogoutRequest;
 import com.yanolja.areas.auth.dto.RegisterRequest;
@@ -26,26 +26,26 @@ import java.util.Map;
 @Tag(name = "인증", description = "인증 관련 API")
 public class AuthController {
 
-    private final AuthService userService;
+    private final AuthService authService;
 
     @Operation(summary = "사용자 회원가입", description = "requestDto[RegisterRequest], responseDto[User]", tags = {"인증"})
     @PostMapping("/register")
     public ApiResponse<User> register(@Valid @RequestBody RegisterRequest registerRequest) {
-        User user = userService.registerUser(registerRequest);
+        User user = authService.registerUser(registerRequest);
         return ApiResponse.success(user, "회원가입이 성공적으로 완료되었습니다.");
     }
 
     @Operation(summary = "사용자 로그인", description = "requestDto[LoginRequest], responseDto[TokenResponse]", tags = {"인증"})
     @PostMapping("/login")
     public ApiResponse<TokenResponse> login(@Valid @RequestBody LoginRequest loginRequest) {
-        TokenResponse tokenResponse = userService.login(loginRequest);
+        TokenResponse tokenResponse = authService.login(loginRequest);
         return ApiResponse.success(tokenResponse, "로그인이 성공적으로 완료되었습니다.");
     }
     
     @Operation(summary = "사용자 로그아웃", description = "requestDto[LogoutRequest], responseDto[Map<String, Object>]", tags = {"인증"})
     @PostMapping("/logout")
     public ApiResponse<Map<String, Object>> logout(@Valid @RequestBody LogoutRequest logoutRequest) {
-        boolean result = userService.logout(logoutRequest);
+        boolean result = authService.logout(logoutRequest);
         Map<String, Object> data = new HashMap<>();
         data.put("success", result);
         String message = result ? "로그아웃 되었습니다." : "로그아웃 처리 중 오류가 발생했습니다.";
@@ -55,7 +55,7 @@ public class AuthController {
     @Operation(summary = "토큰 갱신", description = "requestDto[TokenRefreshRequest], responseDto[TokenResponse]", tags = {"인증"})
     @PostMapping("/refresh")
     public ApiResponse<TokenResponse> refreshToken(@Valid @RequestBody TokenRefreshRequest refreshRequest) {
-        TokenResponse tokenResponse = userService.refreshToken(refreshRequest);
+        TokenResponse tokenResponse = authService.refreshToken(refreshRequest);
         return ApiResponse.success(tokenResponse, "토큰이 성공적으로 갱신되었습니다.");
     }
 }
