+++ src/main/java/com/yanolja/areas/auth/controller/OAuth2Controller.java
@@ -1,6 +1,6 @@
 package com.yanolja.areas.auth.controller;
 
-import com.yanolja.areas.auth.domain.SocialProvider;
+import com.yanolja.areas.user.domain.SocialProvider;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
